refactor: simplify remote content handling
